backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
httpx>=0.26.0
docker>=7.0.0
proxmoxer>=2.0.0
apscheduler>=3.10.0
qdrant-client>=1.7.0
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp-proto-grpc>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0
opentelemetry-instrumentation-httpx>=0.47b0
opentelemetry-instrumentation-sqlalchemy>=0.47b0
PyYAML>=6.0.1
jinja2>=3.1.0
importlib-metadata>=6.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
